+++ src/actions/update_user.rs
@@ -16,7 +16,7 @@ impl<U: UserRepository> UpdateUserAction<U> {
     )]
     pub async fn execute(
         &self,
-        user_id: i32,
+        user_id: u64,
         name: &str,
         email: &str,
     ) -> Result<AuthUser, AuthError> {
